refactor(boot): split RuntimeAutoConfiguration into focused @Configuration classes
Decomposes the ~300-line monolith into three focused top-level classes
(A2aInfrastructureConfiguration, RuntimeLifecycleConfiguration,
A2aExecutionConfiguration), all flat in com.huawei.ascend.runtime.boot.
RuntimeAutoConfiguration becomes a thin @Import composer with no @Bean
methods except a2aPushSender (kept here so existing unit tests that
instantiate the class directly remain callable without a Spring context).
A2aServerExecutor stays nested in RuntimeAutoConfiguration so tests that
reference RuntimeAutoConfiguration.A2aServerExecutor compile unchanged.
The static toTrajectorySettings helpers stay on RuntimeAutoConfiguration
for the same reason. Every bean name, return type, and condition is
preserved exactly; the .imports file is byte-identical.
